RegistrationManagementPage splitted into Components
